Fix issues from merge of reshape-pipe-experiments
Renamed XyTransformer to NpXyTransformer. That is because that class
only accepts NumPy arrays in the initializier anyway. It seems we may
need to convert from other types (say, from a xarray.Dataset back to a
numpy.ndarray), so better make it clear that this is not a "generic
transformer". We may want to think about the design of the various
transformers we will need.

We no longer have a standalone function astype. That was not really
standalone anyway. It called NpXyTransformer, and it didn't quite work
correctly outside of that setting, so I put it back inside
NpXyTransformer.

We also no longer have the TypeTransformerBase class. The only other
class that inherited from it was MLDataset, and just so it could have a
astype method. But I don't expect that transform would work anyway (from
xarray.Dataset to something else; NpXyTransformer.astype is from numpy
arrays to something else).

The NpXyTransformer class is back at datasets.py (where the old
XyTransformer class was).

With those changes, there is no code in xarray_filters/astype.py
anymore.

However, I think we should move NpXyTransformer to astype.com (and any
"converters"). I didn't move it now because I edited the docs of
datasets.py as a whole, and I didn't want to reedit them now (doctests
for datasets.py pass under Python 3right now).

Besides that, we replaced XyTransformer with NpXyTransformer in the
examples/2017*.ipynb notebook. After the NpXyTransformer.to_dataset
method gets some Examples/doctests, I think we should remove this
notebook.
